+++ build.dart
@@ -49,6 +49,7 @@ const scripts = {
   'add_new_tab_button': '*://online.manchester.ac.uk/ultra/*',
   'add_course_images': '*://online.manchester.ac.uk/ultra/*',
   'auto_login': '*://login.manchester.ac.uk/cas/login*',
+  'expand_menu': '*://online.manchester.ac.uk/webapps/blackboard/content/*',
   'video_keyboard_shortcuts': '*://video.manchester.ac.uk/embedded/*',
 };